Update cppcheck support to 1.76.1
Suppressed some classes of warnings where cppcheck was systematically
failing (for at least some source files). Added assert statements
to cover others.

Propagated some const correctness.

Replaced some block_* macros with templated functions in a header
that we re-use in a few places.

The gmx_arpack and lapack issues were inherited from the original
FORTRAN.

The simd/support.cpp issue is a bug that is fixed in release-2016.

Removed some unnecessary stderr output.

Noted TODOs for further cleanup.

+++ src/gromacs/commandline/pargs.cpp
@@ -210,7 +210,7 @@ int getDefaultXvgFormat(gmx::ConstArrayRef<const char *> xvgFormats)
             std::find(xvgFormats.begin(), xvgFormats.end(), std::string(select));
         if (i != xvgFormats.end())
         {
-            return i - xvgFormats.begin();
+            return static_cast<int>(i - xvgFormats.begin());
         }
         else
         {
